# openapi.yaml
openapi: 3.0.3
info:
  title: 千界花园 (Thousand Realms Garden) API
  description: '群智协同操作系统 REST API — XYZ 三轴自由连接、多 Agent 编排、统一 LLM 适配


    📊 自动扩展: 145 个新端点已从后端路由同步'
  version: 1.0.0
  contact:
    name: Thousand Realms Garden Team
  license:
    name: MIT
servers:
- url: http://localhost:3001/api
  description: 本地开发环境
- url: https://thousand-realms-garden.onrender.com/api
  description: 生产环境 (Render)
...
paths:
...
  /events:
    get:
      tags:
      - Events
      summary: GET /events
      operationId: get_events
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
    post:
      tags:
      - Events
      summary: POST /events
      operationId: post_events
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /registry:
    get:
      tags:
      - Registry
      summary: GET /registry
      operationId: get_registry
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
  /registry/axis/{x}/{y}/{z}:
    get:
      tags:
      - Registry
      summary: GET /registry/axis/{x}/{y}/{z}
      operationId: get_registry_axis_x_y_z
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
  /registry/search:
    get:
      tags:
      - Registry
      summary: GET /registry/search
      operationId: get_registry_search
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
  /registry/stats/overview:
    get:
      tags:
      - Registry
      summary: GET /registry/stats/overview
      operationId: get_registry_stats_overview
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
  /registry/{id}:
    get:
      tags:
      - Registry
      summary: GET /registry/{id}
      operationId: get_registry_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /registry/{id}/heartbeat:
    post:
      tags:
      - Registry
      summary: POST /registry/{id}/heartbeat
      operationId: post_registry_id_heartbeat
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /security/audit-log:
    get:
      tags:
      - Security
      summary: GET /security/audit-log
      operationId: get_security_audit-log
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
  /security/blocked-ips:
    get:
      tags:
      - Security
      summary: GET /security/blocked-ips
      operationId: get_security_blocked-ips
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
    post:
      tags:
      - Security
      summary: POST /security/blocked-ips
      operationId: post_security_blocked-ips
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /security/blocked-ips/{ip}:
    delete:
      tags:
      - Security
      summary: DELETE /security/blocked-ips/{ip}
      operationId: delete_security_blocked-ips_ip
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: ip
        in: path
        required: true
        schema:
          type: string
  /security/events:
    get:
      tags:
      - Security
      summary: GET /security/events
      operationId: get_security_events
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
  /security/events/{id}/resolve:
    patch:
      tags:
      - Security
      summary: PATCH /security/events/{id}/resolve
      operationId: patch_security_events_id_resolve
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /security/overview:
    get:
      tags:
      - Security
      summary: GET /security/overview
      operationId: get_security_overview
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 通过 /auth/login 获取的 JWT Token